backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
sqlalchemy-utils==0.41.1

# LLM utilities
requests==2.31.0

# Apple MLX framework for local LLM inference
mlx>=0.0.5
mlx-lm>=0.0.3
